# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: School Management System API
  description: |
    Multi-tenant School Management System REST API

    ## Features
    - JWT Authentication
    - Role-based Access Control
    - Multi-tenant Architecture
    - Comprehensive School Management

    ## Roles
    - **Super Admin**: System-level access, can create schools
    - **School Admin**: School-level access, can manage teachers, students, and school operations
    - **Accountant**: Financial management and limited academic operations
    - **Teacher**: Academic operations, attendance, and syllabus tracking
    - **Parent**: Read-only access to child's information
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.schoolmanagement.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "AUTH_001"
            message: "Unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "AUTH_003"
            message: "Forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "NOT_FOUND"
            message: "Resource not found"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            errorCode: "VALIDATION_ERROR"
            message: "Validation failed"
            errors:
              - field: "email"
                message: "Invalid email format"
...
  schemas:
    Error:
      type: object
      required:
        - success
        - errorCode
        - message
      properties:
        success:
          type: boolean
          example: false
        errorCode:
          type: string
          example: "AUTH_001"
        message:
          type: string
          example: "Invalid email or password"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    Success:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
...
    Student:
      type: object
      properties:
        id:
          type: integer
          example: 1
        schoolId:
          type: integer
          example: 1
        admissionNumber:
          type: string
          example: "2024001"
        fullName:
          type: string
          example: "John Doe"
        dateOfBirth:
          type: string
          format: date
          nullable: true
          example: "2010-05-15"
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
          example: "male"
        address:
          type: string
          nullable: true
          example: "123 Main Street"
        phone:
          type: string
          nullable: true
          example: "1234567890"
        admissionDate:
          type: string
          format: date
          example: "2024-04-01"
        currentClassId:
          type: integer
          nullable: true
          example: 1
        className:
          type: string
          nullable: true
          example: "Class 5"
        classSection:
          type: string
          nullable: true
          example: "A"
        parentId:
          type: integer
          nullable: true
          example: 5
        rollNumber:
          type: string
          nullable: true
          example: "01"
        status:
          type: string
          enum: [active, inactive, transferred, graduated]
          example: "active"
        createdAt:
          type: string
          format: date-time
          example: "2024-04-01T10:30:00.000Z"
    CreateStudentRequest:
      type: object
      required:
        - admissionNumber
        - fullName
        - admissionDate
      properties:
        admissionNumber:
          type: string
          minLength: 1
          maxLength: 50
          example: "2024001"
        fullName:
          type: string
          minLength: 3
          maxLength: 150
          example: "John Doe"
        dateOfBirth:
          type: string
          format: date
          example: "2010-05-15"
        gender:
          type: string
          enum: [male, female, other]
          example: "male"
        address:
          type: string
          example: "123 Main Street"
        phone:
          type: string
          example: "1234567890"
        admissionDate:
          type: string
          format: date
          example: "2024-04-01"
        currentClassId:
          type: integer
          example: 1
        parentId:
          type: integer
          example: 5
        rollNumber:
          type: string
          maxLength: 20
          example: "01"
        status:
          type: string
          enum: [active, inactive, transferred, graduated]
          default: "active"
          example: "active"
    UpdateStudentRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 3
          maxLength: 150
          example: "John Doe"
        dateOfBirth:
          type: string
          format: date
          example: "2010-05-15"
        gender:
          type: string
          enum: [male, female, other]
          example: "male"
        address:
          type: string
          example: "123 Main Street"
        phone:
          type: string
          example: "1234567890"
        currentClassId:
          type: integer
          example: 2
        parentId:
          type: integer
          example: 5
        rollNumber:
          type: string
          maxLength: 20
          example: "15"
        status:
          type: string
          enum: [active, inactive, transferred, graduated]
          example: "active"

    # Subject related schemas
    Subject:
      type: object
      properties:
        id:
          type: integer
          example: 1
        schoolId:
          type: integer
          example: 1
        name:
          type: string
          example: "Mathematics"
        code:
          type: string
          example: "MATH"
        chapterCount:
          type: integer
          example: 12
    CreateSubjectRequest:
      type: object
      required:
        - name
        - code
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Mathematics"
        code:
          type: string
          minLength: 2
          maxLength: 20
          example: "MATH"
    UpdateSubjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Advanced Mathematics"
        code:
          type: string
          minLength: 2
          maxLength: 20
          example: "AMATH"

    # Chapter related schemas
...
paths:
...
  /students:
    post:
      tags:
        - Students
      summary: Create student
      description: Create a new student (School Admin or Accountant)
      operationId: createStudent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
            examples:
              completeStudent:
                summary: Create student with all details
                value:
                  admissionNumber: "STU2024001"
                  fullName: "John Doe"
                  dateOfBirth: "2010-05-15"
                  gender: "male"
                  address: "123 Main Street, City"
                  phone: "+1234567890"
                  admissionDate: "2024-04-01"
                  currentClassId: 5
                  parentId: 10
                  rollNumber: "25"
              basicStudent:
                summary: Create student with required fields only
                value:
                  admissionNumber: "STU2024002"
                  fullName: "Jane Smith"
                  admissionDate: "2024-04-01"
      responses:
        "201":
          description: Student created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Student"
              example:
                success: true
                message: "Student created successfully"
                data:
                  id: 15
                  schoolId: 1
                  admissionNumber: "STU2024001"
                  fullName: "John Doe"
                  dateOfBirth: "2010-05-15"
                  gender: "male"
                  address: "123 Main Street, City"
                  phone: "+1234567890"
                  admissionDate: "2024-04-01"
                  currentClassId: 5
                  className: "Class 5"
                  classSection: "A"
                  parentId: 10
                  rollNumber: "25"
                  status: "active"
                  createdAt: "2024-04-01T10:30:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      tags:
        - Students
      summary: Get all students
      description: Retrieve list of all students in the school
      operationId: getAllStudents
      security:
        - BearerAuth: []
      parameters:
        - name: classId
          in: query
          description: Filter by class
          required: false
          schema:
            type: integer
          example: 5
        - name: status
          in: query
          description: Filter by status
          required: false
          schema:
            type: string
            enum: [active, inactive, transferred, graduated]
          example: "active"
      responses:
        "200":
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Student"
              example:
                success: true
                message: "Students retrieved successfully"
                data:
                  - id: 15
                    schoolId: 1
                    admissionNumber: "STU2024001"
                    fullName: "John Doe"
                    dateOfBirth: "2010-05-15"
                    gender: "male"
                    currentClassId: 5
                    className: "Class 5"
                    classSection: "A"
                    rollNumber: "25"
                    status: "active"
                  - id: 16
                    schoolId: 1
                    admissionNumber: "STU2024002"
                    fullName: "Jane Smith"
                    dateOfBirth: "2010-08-20"
                    gender: "female"
                    currentClassId: 5
                    className: "Class 5"
                    classSection: "A"
                    rollNumber: "26"
                    status: "active"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /students/{id}:
    get:
      tags:
        - Students
      summary: Get student by ID
      description: Retrieve student details by ID
      operationId: getStudentById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 15
      responses:
        "200":
          description: Student retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Student"
              example:
                success: true
                message: "Student retrieved successfully"
                data:
                  id: 15
                  schoolId: 1
                  admissionNumber: "STU2024001"
                  fullName: "John Doe"
                  dateOfBirth: "2010-05-15"
                  gender: "male"
                  address: "123 Main Street, City"
                  phone: "+1234567890"
                  admissionDate: "2024-04-01"
                  currentClassId: 5
                  className: "Class 5"
                  classSection: "A"
                  parentId: 10
                  rollNumber: "25"
                  status: "active"
                  createdAt: "2024-04-01T10:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - Students
      summary: Update student
      description: Update student details (School Admin or Accountant)
      operationId: updateStudent
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 15
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentRequest"
            examples:
              updateClass:
                summary: Update student class
                value:
                  currentClassId: 6
                  rollNumber: "12"
              updateContact:
                summary: Update contact information
                value:
                  phone: "+9876543210"
                  address: "456 New Street, City"
              updateStatus:
                summary: Update student status
                value:
                  status: "inactive"
      responses:
        "200":
          description: Student updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Student"
              example:
                success: true
                message: "Student updated successfully"
                data:
                  id: 15
                  schoolId: 1
                  admissionNumber: "STU2024001"
                  fullName: "John Doe"
                  dateOfBirth: "2010-05-15"
                  gender: "male"
                  address: "123 Main Street, City"
                  phone: "+1234567890"
                  admissionDate: "2024-04-01"
                  currentClassId: 6
                  className: "Class 6"
                  classSection: "A"
                  parentId: 10
                  rollNumber: "12"
                  status: "active"
                  createdAt: "2024-04-01T10:30:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Students
      summary: Deactivate student
      description: Deactivate a student (School Admin only)
      operationId: deleteStudent
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 15
      responses:
        "200":
          description: Student deactivated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
              example:
                success: true
                message: "Student deactivated successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # Subjects endpoints
  /subjects:
    post:
      tags:
        - Subjects
      summary: Create subject
      description: Create a new subject (School Admin or Accountant)
      operationId: createSubject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectRequest"
            examples:
              scienceSubject:
                summary: Create Science subject
                value:
                  name: "Science"
                  code: "SCI"
              mathSubject:
                summary: Create Mathematics subject
                value:
                  name: "Mathematics"
                  code: "MATH"
      responses:
        "201":
          description: Subject created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Subject"
              example:
                success: true
                message: "Subject created successfully"
                data:
                  id: 1
                  schoolId: 1
                  name: "Science"
                  code: "SCI"
                  createdAt: "2024-04-01T10:00:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    get:
      tags:
        - Subjects
      summary: Get all subjects
      description: Retrieve list of all subjects in the school
      operationId: getAllSubjects
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Subjects retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Subject"
              example:
                success: true
                message: "Subjects retrieved successfully"
                data:
                  - id: 1
                    schoolId: 1
                    name: "Science"
                    code: "SCI"
                    createdAt: "2024-04-01T10:00:00Z"
                  - id: 2
                    schoolId: 1
                    name: "Mathematics"
                    code: "MATH"
                    createdAt: "2024-04-01T10:05:00Z"
                  - id: 3
                    schoolId: 1
                    name: "English"
                    code: "ENG"
                    createdAt: "2024-04-01T10:10:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /subjects/{id}:
    get:
      tags:
        - Subjects
      summary: Get subject by ID
      description: Retrieve subject details by ID
      operationId: getSubjectById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Subject retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Subject"
              example:
                success: true
                message: "Subject retrieved successfully"
                data:
                  id: 1
                  schoolId: 1
                  name: "Science"
                  code: "SCI"
                  createdAt: "2024-04-01T10:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - Subjects
      summary: Update subject
      description: Update subject details (School Admin or Accountant)
      operationId: updateSubject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubjectRequest"
            examples:
              updateName:
                summary: Update subject name
                value:
                  name: "Advanced Science"
              updateCode:
                summary: Update subject code
                value:
                  code: "ASCI"
              updateBoth:
                summary: Update name and code
                value:
                  name: "Physical Science"
                  code: "PHYS"
      responses:
        "200":
          description: Subject updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Subject"
              example:
                success: true
                message: "Subject updated successfully"
                data:
                  id: 1
                  schoolId: 1
                  name: "Advanced Science"
                  code: "SCI"
                  createdAt: "2024-04-01T10:00:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - Subjects
      summary: Delete subject
      description: Delete a subject (School Admin only)
      operationId: deleteSubject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Subject deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
              example:
                success: true
                message: "Subject deleted successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # Chapters endpoints
...